Guess Forban name if not set in config
The Forban name is not linked to the UUID and is there
to help users to name/search for a device. So you can
change the forban name when the user wants. This is
also useful for self-deployable Forban where there is
no name set by default and guessing Forban name is
usually enough to get a proper name. This should
work on Windows and Unix-like platform.